requirements.txt
# DeepSeek WhatsApp Bot Dependencies

# Core dependencies
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0

# Logging
loguru==0.7.2 